+++ src/layers/multi_head_attention.cpp
@@ -1,7 +1,7 @@
 #include <layers/multi_head_attention.h>
 
 
-tfm::MultiHeadAttention::MultiHeadAttention(int num_heads, int d_model, std::string filename) :
+tfm::MultiHeadAttention::MultiHeadAttention(size_t num_heads, size_t d_model, std::string filename) :
 	num_heads(num_heads),
 	d_model(d_model),
 	output_(),
